Research_AI_law/F5_F8_MVP/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "research-ai-law-mvp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "matplotlib>=3.7",
  "seaborn>=0.13",
  "pyyaml>=6.0",
  "openpyxl>=3.1",
  "xlsxwriter>=3.1",
  "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["_common*", "FASE5*"]

[tool.pytest.ini_options]
testpaths = ["FASE5/tests"]
addopts = "-q"
